+++ core/src/main/kotlin/sift/core/dsl/Template.kt
@@ -37,42 +37,85 @@ class Template internal constructor(
     }
 
     /** iterates all classes */
-    fun classes(f: Classes.() -> Unit) {
-        action += Action.Fork(
+    fun classes(f: Classes.() -> Unit) = classes(null, f)
+
+    /** iterates all classes */
+    fun classes(label: String?, f: Classes.() -> Unit) {
+        action += Action.Fork(label,
             Action.Template.InstrumentClasses andThen Classes().also(f).action
         )
     }
 
     /** iterates class elements of registered [entity] type */
-    fun classesOf(entity: Entity.Type, f: Classes.(Entity.Type) -> Unit) {
+    fun classesOf(
+        entity: Entity.Type,
+        f: Classes.(Entity.Type) -> Unit
+    ) = classesOf(null, entity, f)
+
+    /** iterates class elements of registered [entity] type */
+    fun classesOf(
+        label: String?,
+        entity: Entity.Type,
+        f: Classes.(Entity.Type) -> Unit
+    ) {
         val classes = Action.Template.ClassesOf(entity)
         val forkTo = Classes().apply { f(entity) }.action
 
-        action += Action.Fork(classes andThen forkTo)
+        action += Action.Fork(label, classes andThen forkTo)
     }
 
     /** iterates method elements of registered [entity] type */
-    fun methodsOf(entity: Entity.Type, f: Methods.(Entity.Type) -> Unit) {
+    fun methodsOf(
+        entity: Entity.Type,
+        f: Methods.(Entity.Type) -> Unit
+    ) = methodsOf(null, entity, f)
+
+    /** iterates method elements of registered [entity] type */
+    fun methodsOf(
+        label: String?,
+        entity: Entity.Type,
+        f: Methods.(Entity.Type) -> Unit
+    ) {
         val methods = Action.Template.MethodsOf(entity)
         val forkTo = Methods().apply { f(entity) }.action
 
-        action += Action.Fork(methods andThen forkTo)
+        action += Action.Fork(label, methods andThen forkTo)
     }
 
     /** iterates field elements of registered [entity] type */
-    fun fieldsOf(entity: Entity.Type, f: Fields.(Entity.Type) -> Unit) {
+    fun fieldsOf(
+        entity: Entity.Type,
+        f: Fields.(Entity.Type) -> Unit
+    ) = fieldsOf(null, entity, f)
+
+    /** iterates field elements of registered [entity] type */
+    fun fieldsOf(
+        label: String?,
+        entity: Entity.Type,
+        f: Fields.(Entity.Type) -> Unit
+    ) {
         val fields = Action.Template.FieldsOf(entity)
         val forkTo = Fields().apply { f(entity) }.action
 
-        action += Action.Fork(fields andThen forkTo)
+        action += Action.Fork(label, fields andThen forkTo)
     }
 
     /** iterates "scope-erased" elements, useful for property tagging entities. */
-    fun elementsOf(entity: Entity.Type, f: Elements.(Entity.Type) -> Unit) {
+    fun elementsOf(
+        entity: Entity.Type,
+        f: Elements.(Entity.Type) -> Unit
+    ) = elementsOf(null, entity, f)
+
+    /** iterates "scope-erased" elements, useful for property tagging entities. */
+    fun elementsOf(
+        label: String?,
+        entity: Entity.Type,
+        f: Elements.(Entity.Type) -> Unit
+    ) {
         val elements = Action.Template.ElementsOf(entity)
         val forkTo = Elements().apply { f(entity) }.action
 
-        action += Action.Fork(elements andThen forkTo)
+        action += Action.Fork(label, elements andThen forkTo)
     }
 
     /** associates all entities with [rhs] */